requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.40.0
django-ckeditor==5.8.0
django-fcm==0.1.1
django-filter==2.2.0
django-js-asset==1.2.2
django-kronos==1.0
django-rest-auth==0.9.5
django-uuidfield==0.5.0
djangorestframework==3.10.3
django-cors-headers==3.2.1
drf-yasg==1.17.0
fcm-django==0.3.2
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
kombu==3.0.37
Markdown==3.1.1
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==19.2
Pillow==6.2.1
psycopg2-binary==2.8.4
pyfcm==1.4.7
pyparsing==2.4.5
python-crontab==2.4.0
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
sendgrid==6.1.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
whitenoise==4.1.4